+++ src/main/java/calculator/tokenizer/SpaceExpressionTokenizer.java
@@ -1,5 +1,6 @@
-package calculator;
+package calculator.tokenizer;
 
+import calculator.tokenizer.ExpressionTokenizer;
 import java.util.Arrays;
 import java.util.List;
