+++ src/test/java/calculator/ExpressionParserTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 
 import calculator.expression.Expression;
 import calculator.expression.NumberExpression;
+import calculator.tokenizer.SpaceExpressionTokenizer;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
